+++ src/Service/PostcodeService.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace JustSteveKing\LaravelPostcodes\Service;
 
 use GuzzleHttp\Client;
+use Illuminate\Support\Collection;
+use function GuzzleHttp\Psr7\build_query;
 use GuzzleHttp\Exception\GuzzleException;
 use JustSteveKing\LaravelPostcodes\Service\BulkReverseGeocoding\Geolocation;
 
@@ -59,6 +61,31 @@ class PostcodeService
     }
 
     /**
+     * Get the address details from a multiple postcodes at once
+     *
+     * @param array $postcodes
+     *
+     * @param array $filter - optional array of fields to return
+     * @return Collection
+     *
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     */
+    public function getPostcodes(array $postcodes, array $filter = []): Collection
+    {
+        if (!empty($filter)) {
+            $filter = build_query(['filter' => implode(',', $filter)]);
+        }
+
+        return collect($this->getResponse(
+            'postcodes?' . $filter,
+            'POST',
+            ['postcodes' => array_values($postcodes)]
+        ))->map(function ($item) {
+            return $item->result;
+        });
+    }
+
+    /**
      * Get information based on outward code including geo data
      *
      * @param string $outwardcode
@@ -83,15 +110,17 @@ class PostcodeService
     /**
      * Query the API for a given string
      *
-     * @param  string  $query
+     * @param string $query
      *
-     * @return array|null
+     * @return Collection
+     *
+     * @throws \GuzzleHttp\Exception\GuzzleException
      */
-    public function query(string $query): ?array
+    public function query(string $query): Collection
     {
         $queryString = http_build_query(['q' => $query]);
 
-        return $this->getResponse("postcodes?$queryString");
+        return collect($this->getResponse("postcodes?$queryString"));
     }
 
     /**
@@ -99,11 +128,13 @@ class PostcodeService
      *
      * @param string $postcode
      *
-     * return array|null
+     * @return Collection
+     *
+     * @throws \GuzzleHttp\Exception\GuzzleException
      */
-    public function nearest(string $postcode): ?array
+    public function nearest(string $postcode): Collection
     {
-        return $this->getResponse("postcodes/$postcode/nearest");
+        return collect($this->getResponse("postcodes/$postcode/nearest"));
     }
 
     /**
@@ -123,28 +154,75 @@ class PostcodeService
      *
      * @param string $partialPostcode
      *
-     * @return array|null
+     * @return Collection
+     *
+     * @throws \GuzzleHttp\Exception\GuzzleException
      */
-    public function autocomplete(string $partialPostcode): ?array
+    public function autocomplete(string $partialPostcode): Collection
     {
-        return $this->getResponse("postcodes/$partialPostcode/autocomplete");
+        return collect($this->getResponse("postcodes/$partialPostcode/autocomplete"));
     }
 
     /**
      * Get nearest outward codes for a given longitude & latitude
      *
-     * @param float $latitude
      * @param float $longitude
      *
-     * @return array|null
+     * @param float $latitude
+     * @return Collection
+     *
+     * @throws \GuzzleHttp\Exception\GuzzleException
      */
-    public function nearestOutwardCodesForGivenLngAndLat(float $longitude, float $latitude): ?array
+    public function nearestOutwardCodesForGivenLngAndLat(float $longitude, float $latitude): Collection
     {
-        return $this->getResponse(sprintf(
+        return collect($this->getResponse(sprintf(
             'outcodes?lon=%s&lat=%s',
             $longitude,
             $latitude
+        )));
+    }
+
+    /**
+     * Get information about nearest outcodes based on outward code
+     *
+     * @param string $outwardcode
+     *
+     * @param int $limit Needs to be less than 100
+     * @param int $radius Needs to be less than 25,000m
+     * @return Collection
+     *
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     */
+    public function getNearestOutwardCode(
+        string $outwardcode,
+        int $limit = 10,
+        int $radius = 5000
+    ): Collection {
+        $limit = ($limit > 100) ? 100 : $limit;
+        $radius = ($radius > 100) ? 25000 : $radius;
+
+        return collect($this->getResponse(
+            "outcodes/$outwardcode/nearest?limit=$limit&radius=$radius"
         ));
+    }
+
+    /**
+     * Get nearest postcodes for a given longitude & latitude
+     *
+     * @param float $longitude
+     * @param float $latitude
+     *
+     * @return Collection
+     *
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     */
+    public function nearestPostcodesForGivenLngAndLat(float $longitude, float $latitude): Collection
+    {
+        return collect($this->getResponse(sprintf(
+            'postcodes?lon=%s&lat=%s',
+            $longitude,
+            $latitude
+        )));
     }
 
     /**
@@ -158,30 +236,35 @@ class PostcodeService
         $body = json_encode(array_map(function (Geolocation $geolocation) {
             return $geolocation->toArray();
         }, $geolocations));
-        return $this->getResponse('postcodes', 'POST', $body);
+        return $this->getResponse('postcodes', 'POST', [], ['body' => $body]);
     }
 
     /**
      * Get the response and return the result object
      *
-     * @param string $uri
+     * @param string|null $uri
      * @param string $method
-     * @param string|null $body
-     *
+     * @param array $data - data to be sent in post/patch/put request
+     * @param array $options - array of options to be passed to curl, if $data is passed 'json' will be overwritten
      * @return mixed
-     * @throws GuzzleException
+     * @throws \GuzzleHttp\Exception\GuzzleException
      */
     protected function getResponse(
         string $uri = null,
         string $method = 'GET',
-        string $body = null
+        array $data = [],
+        array $options = []
     ) {
         $url = $this->url . $uri;
+
+        if (!empty($data)) {
+            $options['json'] = $data;
+        }
 
         $request = $this->http->request(
             $method,
             $url,
-            ['body' => $body]
+            $options
         );
 
         return json_decode($request->getBody()->getContents())->result;